+++ crates/sdk-core/tests/integ_tests/polling_tests.rs
@@ -32,9 +32,12 @@ use temporalio_common::{
     telemetry::{CoreLogStreamConsumer, Logger, TelemetryOptions},
 };
 use temporalio_macros::{workflow, workflow_methods};
-use temporalio_sdk::{ActivityOptions, WorkflowContext, WorkflowResult};
+use temporalio_sdk::{
+    ActivityOptions, WorkflowContext, WorkflowResult,
+    runtime::{AutoscalingOptions, PollerBehavior},
+};
 use temporalio_sdk_core::{
-    CoreRuntime, PollerBehavior, RuntimeOptions, TunerHolder,
+    CoreRuntime, RuntimeOptions, TunerHolder,
     ephemeral_server::{TemporalDevServerConfig, default_cached_download},
     init_worker, prost_dur,
     test_help::{NAMESPACE, WorkerTestHelpers, drain_pollers_and_shutdown, schedule_activity_cmd},
@@ -275,11 +278,13 @@ async fn small_workflow_slots_and_pollers(#[values(false, true)] use_autoscaling
     let wf_name = "only_one_workflow_slot_and_two_pollers";
     let mut starter = CoreWfStarter::new(wf_name);
     if use_autoscaling {
-        starter.sdk_config.workflow_task_poller_behavior = Some(PollerBehavior::Autoscaling {
-            minimum: 1,
-            maximum: 5,
-            initial: 1,
-        });
+        starter.sdk_config.workflow_task_poller_behavior = Some(PollerBehavior::Autoscaling(
+            AutoscalingOptions::builder()
+                .minimum(1)
+                .maximum(5)
+                .initial(1)
+                .build(),
+        ));
     } else {
         starter.sdk_config.workflow_task_poller_behavior = Some(PollerBehavior::SimpleMaximum(2));
     }